requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.8.0
attrs==25.3.0
Brotli==1.1.0
cassandra-asyncio-driver==0.1.4
cassandra-driver==3.29.2
causal-conv1d @ file:///mnt/c/Users/user/Desktop/Projects/final/microServiceAPI/services/summarization/model_repository/causal_conv1d-1.5.0.post4%2Bcu124torch2.6cxx11abiFALSE-cp311-cp311-linux_x86_64.whl
ccm==3.1.5
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
einops==0.8.1
email_validator==2.2.0
fastapi==0.115.8
fastapi-cli==0.0.7
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.5.1
geomet==0.2.1.post1
gevent==25.5.1
geventhttpclient==2.3.4
greenlet==3.2.3
grpcio==1.67.1
h11==0.14.0
hf-xet==1.1.5
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.33.4
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
Jinja2==3.1.5
libevdev==0.12
mamba-ssm @ file:///mnt/c/Users/user/Desktop/Projects/final/microServiceAPI/services/summarization/model_repository/mamba_ssm-2.2.4%2Bcu12torch2.6cxx11abiFALSE-cp311-cp311-linux_x86_64.whl
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
marshmallow-sqlalchemy==1.4.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.3
mypy_extensions==1.1.0
networkx==3.5
ninja==1.11.1.4
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
nvidia-pytriton==0.5.14
packaging==24.2
pluggy==1.6.0
propcache==0.3.2
protobuf==5.29.5
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pytest==8.4.1
python-dotenv==1.0.1
python-multipart==0.0.20
python-rapidjson==1.20
PyYAML==6.0.2
pyzmq==27.0.0
regex==2024.11.6
requests==2.32.4
rich==13.9.4
rich-toolkit==0.13.2
safetensors==0.5.3
sh==2.2.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
sympy==1.13.1
tokenizers==0.21.2
torch==2.6.0
tqdm==4.67.1
transformers==4.53.2
triton==3.2.0
tritonclient==2.59.0
typer==0.15.1
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.5.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==14.2
wrapt==1.17.2
yarl==1.20.1
zipp==3.23.0
zope.event==5.1
zope.interface==7.2
